update unit tests for purge_user_cache
